last update

+++ libft/ft_atoi.c
@@ -3,19 +3,19 @@
 /*                                                        :::      ::::::::   */
 /*   ft_atoi.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: kchaouki <kchaouki@student.1337.ma>        +#+  +:+       +#+        */
+/*   By: rrhnizar <rrhnizar@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/29 16:48:33 by kchaouki          #+#    #+#             */
-/*   Updated: 2022/10/11 11:52:45 by kchaouki         ###   ########.fr       */
+/*   Updated: 2023/06/06 12:10:02 by rrhnizar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_atoi(const char *str)
+unsigned long	ft_atoi(const char *str)
 {
-	long	sign;
-	long	ret;
+	int				sign;
+	unsigned long	ret;
 	int		i;
 
 	ret = 0;
